requirements.txt
Flask
requests
beautifulsoup4
ebooklib
gunicorn 
fish-audio-sdk
python-dotenv